Code needs cleanup
Code is all over the place. Also made a bunch of things DoNotDestroyOnLoad so there are lots of references and variables I can probably get rid of. A lot of the were only there to hold references during a scene change so I can get back to them but if they all carry over to a new scene then the references won't get broken.